Started misc code cleanup
- Moved validate_filename() from common.utils to common.files
- Started in-lining some attribute documentation
